Working towards automatically selecting the MOZBC file.
CONFIG/autowrf_classlib.py modified to include an option to select it
and requires user to select such a file when changing time period or
domain.

CONFIG/wrfenvvar - unrelated change, fixed a bug occurring when environmental
variables are not already set in the shell. Now behaves as expected.

configure - changed to exist after wrfenvvar if that script failed to
set the environmental variables.