Untested modifications to PREPINPT code to support lat-lon map projections.
Lat-lon projections are unique for WRF in that the WPS namelist expects dx and
dy to be given in degrees, but the WRF namelist wants them in meters. In order
to accomplish this, I decided to have autowrfchem wait until the met data has
been prepared, then read the DX and DY global attributes (which are given in
meters) and insert these into the WRF namelist.

This required some modification of the python program used to manage the namelists,
because it had been previously designed to ensure that changes to the common
options of WRF and WPS were guaranteed to match. This was handled by requiring
the --force-wrf-only flag given when issuing the "mod" or "modify" command
to this program.

PREPINPUT/prepwps then had to be modified to check if a lat-lon projection is in
use, and if so, use ncdump to get the dx and dy global attributes and set their
values in namelist.input.

Finally, PREPINPUT/precheck was modified to include a check that the map
projection is 'lambert' or 'polar' unless the 'met-only' flag is specified.
This ensures that an NEI-compatible projection is used if NEI emissions are used.