Mainly added Chemistry menu, some other fixes.
-- New option to set chem options quickly. Chemistry options can be defined in
the chemlist.txt file found in CONFIG. A new menu option once the namelists are
loaded allows you to set the chemistry to any of those defined in that file. The
chemlist.txt file will be checked for proper syntax, but any issues will only
result in a printed message, not an error. Does check if KPP is enabled if you
request a chemistry marked as requiring KPP.

-- Fix to NEI compatibility check: verifies that kemit=19.