Bugfixes for the case when YACC & FLEX_LIB_DIR are set in the shell already
MODIFIED FILES:
    CONFIG/wrfenvvar - as it stood, if YACC and/or FLEX_LIB_DIR were already set
in your shell, this would fail because of how it verified that they were
successfully set automatically. It now checks if those variables are already
set, and if they are, verifies that YACC points to a valid file and FLEX_LIB_DIR
contains libfl.a. If override is given, then it will use the normal automatic
method without regard to whether or not they are set already.