Bugfixes in runwrf and some other enhancements
MODIFIED FILES:
    PREPINPUT/prepmozbc - skips over empty lines when reading
mozbc-chemopt-list.txt now as well as lines starting in "#".

    autowrfchem_main - flag to run in restart mode shortened from "restart" to
"rst".

    runwrf - 2 bugfixes: rst boolean was initialized as true, so it would always
try to restart, now initialized as false. Second, when restarting the use of the
tempmod functionality of the Python namelist manager was done incorrectly so
that the temp start date didn't actually get used; both tempmod commands needed
to be combined into one.