Modifications to enable using FDDA nudging with input prep.
FDDA nudging allows you to push the model meteorology toward the reanalysis
meteorology that is already used to generate the initial and boundary
conditions. When FDDA nudging is active (grid_fdda=1 in the &fdda section of the
WRF namelist), real.exe will generate an additional file that contains the
necessary information for nudging. Since this increases the time real takes to
run, we don't want that to happen until the last execution of real.exe during
the input preparation.

MODIFIED FILES:
    CONFIG/autowrf_namelist_main.py - the check-wrf-opt and check-wps-opt
command line methods changed so that, if any checked option cannot be found, the
exit code 2 is given rather than a python error. get-{wrf,wps}-opt changed so
that in the same situation the exit code is 2 (rather than 1).

    PREPINPUT/prepmegan - turns off grid_fdda if necessary before running
real.exe.

    PREPINPUT/prepnei_convert - same idea as prepmegan.

    PREPINPUT/prepmozbc - the python program tempmod command before the last
real.exe should restore grid_fdda if it was on, but this now double-checks that
the expected behavior occurs.

    prepinpt - sets FDDA_ON to reflect whether FDDA nudging is requested in the
namelist.