Added the --run-for option for running WRF, it allows you to specify a shorted run time than the namelist.
MODIFIED FILES:
    Tools/datecompare.py - refactored to allow you to add or subtract
time to either date before the comparison.

    autowrfchem_main - passes the --run-for option to runwrf.

    runwrf - if given the --run-for option, it checks if setting that
run time would make the run shorter than specified in the namelist,
starting from a restart file if desired. If so, then the run time is
changed to that specified in the --run-for option. This is intended
mainly for long runs that need to be broken up into separate jobs,
so you can run for e.g. 30 days from the last restart file, restarting
further each time, until the end time in the namelist is reached. Note
that presently there is no mechanism to halt submission if the end was
already reached.