Mainly a bugfix resulting from FDDA and split-wps interaction
MODIFIED FILES:
    CONFIG/autowrf_classlib.py - print calls outside the UI functions
replaced with msg_print(), which can be turned off by setting the module
variable DEBUG_LEVEL to 0. This prevents printing messages that get taken
by the shell when saving output to a variable with the $() notation.

    CONFIG/autowrf_namelist_main.py - for any shell commands, sets DEBUG_LEVEL
for the classlib to 0.

    PREPINPUT/postcheck.py - close netCDF files when done with them.

    PREPINPUT/metgriblist.py - added necessary kludge to handle run time
starting on the 31st of a month, which erroneously generated NARR3d_yyyymm_3131
as an expected file.

    PREPINPUT/prepnei_convert - copies rsl.error.0000 for each of the convert_emiss
runs to PREPLOGS.