A few more fixes to handle untarring NARR files
MODIFIED FILES:
    PREPINPUT/precheck - the call to metgriblist.py modified to not put $look_for_grib in quotes
because it seemed to treat empty quotes ("") as an extra argument and got confused.

    PREPINPUT/prepwps - now copies each tar file to the MetTARs directory before untarring; this
fixes an issue when multiple instances try to access the file at the same time (as far as I could
tell), which happens frequently when running split WPS.

    PREPINPUT/splitwps - the delay before deleted directories reduced (because a full minute is
too long to wait). Also, put a delay between copying the WPS directory and running it; this seemed
to fix an issue with files not being present once while running split WPS.