First step to 2.0.0
The 2.0.0 version will live completely under the 'autowrfchem' package
directory. This commit adds the entry point, autowrfchem_main.py, in
that package that handles argument parsing. The format of some options
has changed to be compatible with argparse.  Currently, this doesn't do
anything except parse arguments.

The old autowrfchem_main shell script has been moved to
autowrfchem_main.sh to make room for a new autowrfchem_main Python
script which calls the entry point directory (avoiding the need for the
"python -m ..." syntax) and will try to activate a virtual environment
that it lives in.