Reading and writing namelists now uses f90nml
This means that internally the values are stored as the
proper types, rather than always being strings, so a number
of changes were necessary to reflect that in the code. The
biggest is that when setting a namelist option, it checks that
the value is of the proper type or converts it if requested.