First implementation of a reinit driver
For some applications, it is common to reinitialize WRF from analysis
meteorology periodically. I've added an option to support that here.

The process is that, during input preparation, new input and boundary
files are prepared for each reinitialized run and placed in subfolders
of the WRF run directory. Then, during execution, the input files from
those subfolders are linked to the main run folder, the namelist is
configured for the sub-run, and the run is executed.

At present, this is untested and only implemented for WRF, not WRF-Chem.
However, in theory, expanding it for WRF-Chem would not require creating
fresh wrfchemi and wrfbiochemi files, since those should not change across
different reinit runs. It would only require running MOZBC for each time
period, in addition to real.exe.