Implemented fix to let reinit and ensemble modes work together
I realized that there was an incompatibility between the ensemble mode and the
reinit mode. The reinit mode relies on being able to reset the namelist to the
different time periods for each sub-run, while the ensemble mode relies on the
namelist it puts in the WRF directory not changing.

The fix implemented here adds a new command line options, '--alt-namelist', that
allows a different permanent namelist to be specified. The ensemble creator now
creates a unique, alternate permanent namelist in each run directory, and adds
the --alt-namelist option pointing to that namelist instead of using the
--no-sync-namelist option.

Other changes:
* The `clear_temp_changes` method on the namelist container now allows you to
 specify which WRF/WPS directory to clear changes in. The `_run_wrf` function
 uses this to ensure that it resets the proper directory, if an alternate WRF
 run directory was given.
* Added `ignore_if_done` option for the ensemble runner, which should skip any
 ensemble member if it's already finished. However, this will not work with the
 reinit mode currently, as it expects wrfout files to be the in WRF run
 directory, not the Reinit subdirectories