Fix for WRF/WPS parent_id values.
In WRF, the parent ID of the coarse domain is usually given as 0.
The WPS documentation says it should be 1. Until I have a chance to
test whether 0 or 1 really matters, I've added an extra set of sync
functions for NamelistContainer that will allow WRF to use 0 and WPS
to use 1.

Two small fixes:
* missing space in help string
* added extra bit of commenting to ensemble config file that explains
 that the path to the submit file and ensemble top dir will, if
 relative, be relative to the top of the AutoWRFChem repo.